reorganised functions and filenames to match consistent order
