+++ memort/free_7.c
@@ -1,0 +1,29 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   free_7.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: pandalaf <pandalaf@student.42wolfsburg.    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/02/10 17:06:54 by pandalaf          #+#    #+#             */
+/*   Updated: 2023/02/10 17:08:16 by pandalaf         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/minirt.h"
+#include <stdlib.h>
+
+//FUNCTION WITHOUT DESCRIPTION
+void	free_cylinder_plane(t_cylinder_cap *t)
+{
+	free_plane(t->cap_plane);
+	free(t);
+}
+
+//FUNCTION WITHOUT DESCRIPTION
+void	free_cylinder_cap(t_cylinder_cap *t)
+{
+	free_vector(t->disc_center_to_point_plane);
+	free_plane(t->cap_plane);
+	free(t);
+}
